refactor(analysis): extract function complexity module
